requirements.txt
aiogram==2.23.1
Ydl==1.3.0
youtube-dl
youtube-search-python
spotipy
yt_dlp
moviepy
python-dotenv
